Refactor orchestrator into modular mixins
